backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.3.0
Werkzeug==2.3.7
requests==2.31.0
httpx==0.23.0
# New dependencies for authentication
Flask-JWT-Extended==4.6.0
bcrypt==4.1.2
# New dependencies for email notifications
Flask-Mail==0.9.1
# New dependencies for API documentation
Flask-RESTX==1.3.0
# New dependencies for analytics
python-dateutil==2.8.2
# Dependencies for PII Redaction
PyMuPDF==1.23.26
python-docx==1.1.0
spacy==3.7.2
transformers
torch
sentence-transformers
celery
redis
unstructured[all-docs]
Pillow
chromadb
